minor update filters
